ISPN-10244 Observability: Split RestMetricsClient from the RestClient interface
* move metrics to /metrics because /rest/v2/metrics was not following the spec
