+++ client/rest-client/src/main/java/org/infinispan/client/rest/RestClient.java
@@ -16,6 +16,7 @@ import org.infinispan.commons.util.Experimental;
  **/
 @Experimental
 public interface RestClient extends Closeable {
+
    @Override
    void close() throws IOException;
 
@@ -64,22 +65,12 @@ public interface RestClient extends Closeable {
     */
    RestRawClient raw();
 
+   RestMetricsClient metrics();
+
    /**
     * Returns the configuration of this {@link RestClient}
     */
    RestClientConfiguration getConfiguration();
-
-   CompletionStage<RestResponse> metrics();
-
-   CompletionStage<RestResponse> metrics(String path);
-
-   CompletionStage<RestResponse> metrics(boolean openMetrics);
-
-   CompletionStage<RestResponse> metrics(String path, boolean openMetrics);
-
-   CompletionStage<RestResponse> metricsMetadata();
-
-   CompletionStage<RestResponse> metricsMetadata(String path);
 
    /**
     * Creates a {@link RestClient} instance based on the supplied configuration
