+++ client/rest-client/src/main/java/org/infinispan/client/rest/impl/okhttp/RestClientOkHttp.java
@@ -1,8 +1,5 @@
 package org.infinispan.client.rest.impl.okhttp;
 
-import static org.infinispan.commons.dataconversion.MediaType.APPLICATION_JSON_TYPE;
-import static org.infinispan.commons.dataconversion.MediaType.APPLICATION_OPENMETRICS_TYPE;
-
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -23,6 +20,7 @@ import org.infinispan.client.rest.RestCacheManagerClient;
 import org.infinispan.client.rest.RestClient;
 import org.infinispan.client.rest.RestClusterClient;
 import org.infinispan.client.rest.RestCounterClient;
+import org.infinispan.client.rest.RestMetricsClient;
 import org.infinispan.client.rest.RestRawClient;
 import org.infinispan.client.rest.RestResponse;
 import org.infinispan.client.rest.RestServerClient;
@@ -57,7 +55,6 @@ public class RestClientOkHttp implements RestClient {
    private final RestClientConfiguration configuration;
    private final OkHttpClient httpClient;
    private final String baseURL;
-   private final String baseMetricsURL;
 
    public RestClientOkHttp(RestClientConfiguration configuration) {
       this.configuration = configuration;
@@ -115,7 +112,6 @@ public class RestClientOkHttp implements RestClient {
       httpClient = builder.build();
       ServerConfiguration server = configuration.servers().get(0);
       baseURL = String.format("%s://%s:%d", sslContext == null ? "http" : "https", server.host(), server.port()).replaceAll("//", "/");
-      baseMetricsURL = String.format("%s%s/v2/metrics", baseURL, configuration.contextPath()).replaceAll("//", "/");
    }
 
    @Override
@@ -198,45 +194,8 @@ public class RestClientOkHttp implements RestClient {
    }
 
    @Override
-   public CompletionStage<RestResponse> metrics() {
-      return metricsGet("", false);
-   }
-
-   @Override
-   public CompletionStage<RestResponse> metrics(String path) {
-      return metricsGet(path, false);
-   }
-
-   @Override
-   public CompletionStage<RestResponse> metrics(boolean openMetricsFormat) {
-      return metricsGet("", openMetricsFormat);
-   }
-
-   @Override
-   public CompletionStage<RestResponse> metrics(String path, boolean openMetricsFormat) {
-      return metricsGet(path, openMetricsFormat);
-   }
-
-   @Override
-   public CompletionStage<RestResponse> metricsMetadata() {
-      return metricsOptions("");
-   }
-
-   @Override
-   public CompletionStage<RestResponse> metricsMetadata(String path) {
-      return metricsOptions(path);
-   }
-
-   private CompletionStage<RestResponse> metricsGet(String path, boolean openMetricsFormat) {
-      Request.Builder builder = new Request.Builder()
-            .addHeader("ACCEPT", openMetricsFormat ? APPLICATION_OPENMETRICS_TYPE : APPLICATION_JSON_TYPE);
-      return execute(builder, baseURL, configuration.contextPath(), "v2", "metrics", path);
-   }
-
-   private CompletionStage<RestResponse> metricsOptions(String path) {
-      Request.Builder builder = new Request.Builder()
-            .method("OPTIONS", null);
-      return execute(builder, baseURL, configuration.contextPath(), "v2", "metrics", path);
+   public RestMetricsClient metrics() {
+      return new RestMetricsClientOkHttp(this);
    }
 
    CompletionStage<RestResponse> execute(Request.Builder request) {
